Allow buffer to be placed in DRAM

+++ driver/private_include/cam_hal.h
@@ -45,7 +45,7 @@ esp_err_t cam_deinit(void);
  */
 esp_err_t cam_init(const camera_config_t *config);
 
-esp_err_t cam_config(const camera_config_t *config, framesize_t frame_size, uint8_t sensor_pid);
+esp_err_t cam_config(const camera_config_t *config, framesize_t frame_size, uint16_t sensor_pid);
 
 void cam_stop(void);
